+++ include/NeoN/finiteVolume/cellCentred/operators/ddtOperator.hpp
@@ -11,6 +11,9 @@
 #include "NeoN/linearAlgebra/linearSystem.hpp"
 #include "NeoN/linearAlgebra/sparsityPattern.hpp"
 #include "NeoN/finiteVolume/cellCentred/fields/volumeField.hpp"
+#include "NeoN/finiteVolume/cellCentred/fields/surfaceField.hpp"
+#include "NeoN/timeIntegration/ddt/DdtScheme.hpp"
+#include "NeoN/timeIntegration/ddt/BDF1.hpp"
 
 namespace NeoN::finiteVolume::cellCentred
 {
@@ -31,16 +34,27 @@ public:
 
     void implicitOperation(la::LinearSystem<ValueType, localIdx>& ls, scalar, scalar dt) const;
 
-    void read(const Input&) {}
+    void read(const Input&);
 
     const la::SparsityPattern& getSparsityPattern() const { return sparsityPattern_; }
 
     std::string getName() const { return "DdtOperator"; }
 
+    const timeIntegration::DdtScheme& scheme() const
+    {
+        NF_ASSERT(scheme_ != nullptr, "ddt scheme not configured");
+        return *scheme_;
+    }
+
 private:
 
     // NOTE ddtOperator does not have a FactoryClass
     const la::SparsityPattern& sparsityPattern_;
+
+    static timeIntegration::BDF1 DEFAULT_BDF1_SCHEME;
+    const timeIntegration::DdtScheme* scheme_ {&DEFAULT_BDF1_SCHEME};
+
+    mutable bool firstTimeStep_ {true};
 };